2/10 15:00 1차 회의 골자 완성
