+++ workspace/KHmarketPrj/src/com/kh/account/AccountService.java
@@ -1,0 +1,23 @@
+package com.kh.account;
+
+public class AccountService {
+	public void deposit() {
+		// 충전
+	}
+	
+	public void withdraw() {
+		// 인출
+	}
+	
+	public void transfer() {
+		// 송금
+	}
+	
+	public void receive() {
+		// 수취
+	}
+	
+	public void showEvents() {
+		// 계좌 내역 출력
+	}
+}
