+++ workspace/KHmarketPrj/src/com/kh/board/BoardCategorySelect.java
@@ -1,0 +1,28 @@
+package com.kh.board;
+
+// 카테고리 선
+public class BoardCategorySelect {	
+	public void sellProduct() {
+		// 판매글
+	}
+	
+	public void buyProduct() {
+		// 구매글
+	}
+	
+	public void transactRealEstate() {
+		// 부동산 거래글 
+	}
+	
+	public void useTruck() {
+		// 용달, 트럭 글 
+	}
+	
+	public void findPartTime() {
+		// 알바 구인글
+	}
+	
+	public void searchRestaurant() {
+		// 식당 찾기
+	}
+}
