--- workspace/KHmarketPrj/src/com/kh/board/BoardSelect.java
@@ -1,15 +1,0 @@
-package com.kh.board;
-
-public class BoardSelect {	
-	public void sellProduct() {
-		// 물건 팔기
-	}
-	
-	public void findPartTime() {
-		// 알바 구인
-	}
-	
-	public void searchRestaurant() {
-		// 식당 찾기
-	}
-}
