+++ workspace/KHmarketPrj/src/com/kh/board/BoardService.java
@@ -14,6 +14,14 @@ public class BoardService {
 	}
 
 	public void sendMessage() {
-		// 댓글 쓰기		
+		// 댓글 쓰기 / 메세지 송신		
+	}
+	
+	public void receiveMessage() {
+		// 댓글 받기 / 메세지 수신		
+	}
+	
+	public void selectCategory() {
+		// 카테고리 선택 -> BoardSelect와 연관
 	}
 }
