+++ workspace/KHmarketPrj/src/com/kh/center/CenterService.java
@@ -1,0 +1,11 @@
+package com.kh.center;
+
+public class CenterService {
+	public void askQuestion() {
+		// 질문하기
+	}
+	
+	public void answerQuestion() {
+		// 답변하기
+	}
+}
