+++ workspace/KHmarketPrj/src/com/kh/person/UserTransaction.java
@@ -2,27 +2,11 @@ package com.kh.person;
 
 // 직거래, 당근페이(충전, 송금, 수취), 거래내역 출력
 public class UserTransaction {
-	public void directTransact() {
-		// 직거래
+	public void offlineTransact() {
+		// 직거래 (오프라인 거래)
 	}
 	
-	public void deposit() {
-		// 충전
-	}
-	
-	public void withdraw() {
-		// 인출
-	}
-	
-	public void transfer() {
-		// 송금
-	}
-	
-	public void receive() {
-		// 수취
-	}
-	
-	public void showEvents() {
-		// 거래 내역 출력
+	public void onlineTransact() {
+		// 안전거래 (온라인 거래)
 	}
 }
